simple-crm-app/requirements.txt
asgiref==3.7.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-tailwind==0.5.0
Django==5.0
django-admin==2.0.2
django-crispy-forms==2.0
django-debug-toolbar==4.4.6
django-excel-response2==3.0.5
django-htmx==1.14.0
django-htmx-refresh==0.0.7
django-six==1.0.5
excel-base==1.0.4
gunicorn==20.1.0
idna==3.10
isoweek==1.3.3
Pillow==9.5.0
psycopg==3.1.9
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.32.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.9.0
tzlocal==5.0.1
urllib3==2.2.3
whitenoise==6.6.0
xlwt==1.3.0
